refactor: separate controllers
