+++ app/Http/Controllers/Front/ChatController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Front;
 
+use App\Http\Controllers\Controller;
 use App\Models\Chat;
 use Illuminate\Http\Request;
