+++ app/Http/Controllers/Front/ProductController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Front;
 
+use App\Http\Controllers\Controller;
 use App\Models\Product;
 use App\Models\Shop;
 use Illuminate\Http\Request;
